:recycle: Restrict App 에러코드 추가
