+++ basterdz-common/src/main/java/com/dnd/common/exception/ErrorCode.java
@@ -19,7 +19,10 @@ public enum ErrorCode {
 	//Room
 	ROOM_NOT_FOUND("ROOM-01", "존재하지 않는 그룹입니다."),
 	INVALID_INVITE_CODE("ROOM-02", "올바르지 않은 초대코드입니다."),
-	ALREADY_OVER_PERSONNEL("ROOM-03", "정원을 초과하였습니다");
+	ALREADY_OVER_PERSONNEL("ROOM-03", "정원을 초과하였습니다"),
+	UNSUPPORTED_RESTRICT_APP("ROOM-04", "지원하지 않는 앱입니다.");
+
+	//ScreenTIme
 
 	private final String code;
