--- basterdz-domain/src/main/java/com/dnd/domain/room/entity/RestrictApp.java
@@ -1,18 +1,0 @@
-package com.dnd.domain.room.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum RestrictApp {
-    INSTAGRAM("instagram"),
-    YOUTUBE("youtube"),
-    TIKTOK("tiktok"),
-    NETFLIX("netflix")
-    ;
-
-    private final String appName;
-
-    RestrictApp(String appName) {
-        this.appName = appName;
-    }
-}
